# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/public/configs:
    post:
      tags: [Configs]
      summary: Get public app configs by keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigKeysInput"
      responses:
        "200":
          description: Public config dictionary.
...
  /api/v1/build/templates/{templateId}/preview:
    post:
      tags: [Builds]
      summary: Validate template parameters and preview the generated Dockerfile
      parameters:
        - name: templateId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildTemplatePreviewInput" }
      responses:
        "200":
          description: Rendered, immutable build definition preview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildTemplatePreview" }
  /api/v1/build/environment-config:
    get:
      tags: [Builds]
      summary: Get one global, application, or deployment build environment
      parameters:
        - $ref: "#/components/parameters/BuildEnvironmentScope"
        - { name: projectId, in: query, schema: { type: string } }
        - { name: applicationId, in: query, schema: { type: string } }
        - { name: deploymentTargetId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Public values and boolean secret presence. Secret values and references are never returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildEnvironmentConfig" }
    put:
      tags: [Builds]
      summary: Replace one global, application, or deployment build environment
      parameters:
        - $ref: "#/components/parameters/BuildEnvironmentScope"
        - { name: projectId, in: query, schema: { type: string } }
        - { name: applicationId, in: query, schema: { type: string } }
        - { name: deploymentTargetId, in: query, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildEnvironmentConfigInput" }
      responses:
        "200":
          description: Updated build environment with secret presence only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildEnvironmentConfig" }
...
  /api/v1/auth/bootstrap/admin:
    post:
      tags: [Auth]
      summary: Initialize the first platform admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitializeAdminInput"
      responses:
        "201":
          description: Created platform admin and session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid email, password, language, or JSON (`bootstrap.invalid_input` or `request.invalid_json`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: The production bootstrap token is invalid (`bootstrap.token_invalid`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Platform admin already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Production bootstrap is unavailable because `BOOTSTRAP_TOKEN` is not configured (`bootstrap.unavailable`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with a local account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: Login failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/auth/login/resume:
    post:
      tags: [Auth]
      summary: Resume login with a remembered account
      description: Rotates the per-user remember token, creates a new 24-hour session, and refreshes the 30-day remember cookie. Browser cookies are required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeLoginInput"
      responses:
        "200":
          description: Remembered login resumed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: Remember token missing, expired, revoked, or the account is disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/auth/registration/email/code:
    post:
      tags: [Auth]
      summary: Request an email registration verification code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailRegistrationCodeInput" }
      responses:
        "202":
          description: Verification challenge created.
  /api/v1/auth/registration/email:
    post:
      tags: [Auth]
      summary: Complete email registration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailRegistrationInput" }
      responses:
        "201":
          description: Account created and signed in.
  /api/v1/auth/registration/settings:
    get:
      tags: [Auth]
      summary: Get registration and SMTP settings
      responses:
        "200":
          description: Registration settings with the write-only SMTP password omitted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthRegistrationSettings" }
    put:
      tags: [Auth]
      summary: Update registration and SMTP settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthRegistrationSettingsInput" }
      responses:
        "200":
          description: Registration settings updated.
  /api/v1/auth/mfa/status:
    get:
      tags: [Auth]
      summary: Get current user's MFA status
      description: Requires an interactive browser session. Personal access tokens cannot manage or verify MFA.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Current enrollment, policy, and recovery-code status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFAStatus" }
        "401":
          description: Browser session is missing or invalid (`mfa.session_required` or an authentication error).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Personal access tokens cannot access MFA session endpoints (`mfa.session_required`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: MFA status could not be loaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/auth/mfa/recovery-codes:
    post:
      tags: [Auth]
      summary: Regenerate MFA recovery codes
      description: Requires a valid `mfa_manage` assertion. Replaces and invalidates all previous recovery codes; the new plaintext codes are returned only once.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Recovery codes replaced.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFARecoveryCodes" }
        "401":
          description: Browser session is missing or invalid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: MFA management assertion is missing or expired (`mfa_required`), or a personal access token was used (`mfa.session_required`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: MFA is not enabled (`mfa.not_enabled`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Recovery codes could not be generated or persisted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /api/v1/auth/providers:
    get:
      tags: [Auth]
      summary: List auth providers
      responses:
        "200":
          description: Auth provider list.
    post:
      tags: [Auth]
      summary: Create auth provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthProviderInput"
      responses:
        "201":
          description: Created auth provider.
  /api/v1/auth/providers/{providerId}:
    put:
      tags: [Auth]
      summary: Update auth provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthProviderInput"
      responses:
        "200":
          description: Updated auth provider.
...
  /api/v1/auth/oidc/callback:
    get:
      tags: [Auth]
      summary: Complete OIDC callback
      parameters:
        - $ref: "#/components/parameters/OAuthState"
        - $ref: "#/components/parameters/OAuthCode"
      responses:
        "302":
          description: Redirect after OIDC callback.
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna personal access token
  parameters:
    BuildEnvironmentScope:
      name: scope
      in: query
      required: true
      schema: { type: string, enum: [global, application, deployment] }
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    AuthSessionResponse:
      type: object
      properties:
        user:
          type: object
          additionalProperties: true
      required: [user]
...
    ConfigKeysInput:
      type: object
      properties:
        keys:
          type: array
          items: { type: string }
      required: [keys]
...
    BuildTemplatePreviewInput:
      type: object
      properties:
        version: { type: string }
        values:
          type: object
          additionalProperties: { type: string }
      required: [values]
    BuildTemplatePreview:
      type: object
      properties:
        templateId: { type: string }
        version: { type: string }
        values:
          type: object
          additionalProperties: { type: string }
        dockerfile: { type: string }
        checksum: { type: string }
      required: [templateId, version, values, dockerfile, checksum]
...